backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.9
PyPDF2==3.0.1
requests==2.32.3
beautifulsoup4==4.12.3
openai==1.54.3
python-dotenv==1.0.1
pydantic==2.12.5
scikit-learn==1.5.2
joblib==1.4.2
pandas==2.2.2
numpy==1.26.4



